backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0  # Required for EmailStr validation

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0  # PostgreSQL async driver for SQLAlchemy

# Cache & Queue
redis==5.0.1
celery==5.3.4
pika==1.3.2

# AI & ML
numpy<2.0.0  # IMPORTANT: PyTorch/Whisper requires NumPy 1.x
faster-whisper==1.0.3  # 4-5x faster than openai-whisper, better CPU performance
# openai-whisper==20231117  # Replaced by faster-whisper
# torch>=2.2.0  # Not needed for faster-whisper (uses CTranslate2)
# torchvision>=0.17.0  # Not needed for faster-whisper
transformers>=4.37.0
pillow==10.1.0
google-generativeai>=0.3.0
chromadb>=0.4.22
langchain-text-splitters>=0.0.1

# Agent Orchestration
langgraph>=0.0.40
langchain-core>=0.1.0

# Video Processing
yt-dlp>=2024.4.9
ffmpeg-python==0.2.0
moviepy==1.0.3

# API & Services
httpx==0.25.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1  # Updated for chromadb compatibility

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-slugify==8.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
